pyproject.toml
[project]
name = "fastapi-tortoise-aerich-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich[toml]>=0.8.2",
    "fastapi[standard]>=0.115.12",
    "tortoise-orm[asyncpg]>=0.25.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
]

[tool.aerich]
tortoise_orm = "app.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
